chore: rename GameObjRawPrototype.ts to GameObjRaw
