+++ src/kaplay.ts
@@ -84,7 +84,7 @@ import { pos } from "./ecs/components/transform/pos";
 import { rotate } from "./ecs/components/transform/rotate";
 import { scale } from "./ecs/components/transform/scale";
 import { z } from "./ecs/components/transform/z";
-import { KeepFlags } from "./ecs/entity/GameObjRawPrototype";
+import { KeepFlags } from "./ecs/entity/GameObjRaw";
 import { getCollisionSystem } from "./ecs/systems/collision";
 import { KEvent, KEventController, KEventHandler } from "./events/events";
 import {
